[sistema-login-frontend] add SweetAlert for delete alert

+++ sistema-login-frontend/src/app/list-users/list-users.component.ts
@@ -1,7 +1,9 @@
 import { Component, OnInit } from '@angular/core';
 import { Router } from '@angular/router';
+import swal from 'sweetalert2';
 import { UserService } from '../user.service';
 import { Usuario } from '../usuario';
+
 
 @Component({
   selector: 'app-list-users',
@@ -28,10 +30,32 @@ export class ListUsersComponent implements OnInit {
     this.router.navigate(['edit-user',id]);
   }
 
+
+
   deleteUser(id:number){
-    this.userService.deleteUser(id).subscribe( dato => {
-      this.getUsers();
-    });
+    swal({
+      title: "¿Estas seguro que deseas eliminar el usuario?",
+      text: "Confirma si deseas eliminar el usuario",
+      type: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#d33',
+      confirmButtonText: "Si, eliminar",
+      cancelButtonText: "No, no eliminar",
+      confirmButtonClass: "btn btn-success",
+      cancelButtonClass: "btn btn-danger",
+      buttonsStyling: true
+    }).then((result) => { if (result.value){
+      this.userService.deleteUser(id).subscribe( dato => {
+        this.getUsers();
+        swal(
+          "Usuario eliminado",
+          "El usuario ha sido eliminado con exito",
+          'success'
+        );
+      });
+    } 
+  })
+    
   }
 
   userDetails(id:number){
